+++ firmware/hdl/payload/gtl/bx_pipeline.vhd
@@ -2,6 +2,7 @@
 -- Pipeline for +/-2 bx data.
 
 -- Version-history:
+-- HB 2023-07-25: updated for new ZDC data structure.
 -- HB 2022-10-10: updated for ZDC (on link 11).
 -- HB 2021-06-10: updated for hadronic shower trigger (mus).
 -- HB 2021-05-21: first design.
@@ -67,8 +68,8 @@ begin
         bx_data.asymht(i) <= data_tmp(i).asymht;
         bx_data.asymethf(i) <= data_tmp(i).asymethf;
         bx_data.asymhthf(i) <= data_tmp(i).asymhthf;
-        bx_data.zdcp(i) <= data_tmp(i).zdc(0);
-        bx_data.zdcm(i) <= data_tmp(i).zdc(1);
+        bx_data.zdcm(i) <= data_tmp(i).zdc(1)(ZDC_BIT_HIGH downto ZDC_BIT_LOW);
+        bx_data.zdcp(i) <= data_tmp(i).zdc(2)(ZDC_BIT_HIGH downto ZDC_BIT_LOW);
 
 -- Additional delay for centrality and ext_cond (no comparators and conditions)